tests/test_routes.py
"""
Tests for Flask route endpoints.

Uses the Flask test client to verify HTTP responses, authentication handling,
and JSON serialization for key REST API endpoints.
"""

import json
import os
from unittest.mock import MagicMock, patch

import pytest

# ---------------------------------------------------------------------------
# Helpers
# ---------------------------------------------------------------------------


def _patch_validate_decorate():
    """Patch validate_decorate to bypass JWT verification in tests."""
    from flask import g

    def fake_validate(request):
        if request.method != "OPTIONS":
            g.user = "test-user-uuid-1234"
            g.token = "fake-jwt-token"

    return patch("routes.utils.validate_decorate", side_effect=fake_validate)


def _patch_decode_token():
    """Patch decode_token to return a valid decoded JWT."""
    return patch(
        "routes.utils.decode_token",
        return_value={
            "sub": "test-user-uuid-1234",
            "preferred_username": "testuser",
            "resource_access": {
                "quantimage2-frontend": {"roles": ["admin"]},
            },
        },
    )


# ---------------------------------------------------------------------------
# Entrypoint / health check
# ---------------------------------------------------------------------------


class TestEntrypoint:
    def test_root_returns_status(self, client, app):
        """GET / should return a status message."""
        with _patch_validate_decorate(), _patch_decode_token():
            response = client.get("/")

        assert response.status_code == 200
        data = response.get_json()
        assert "status" in data


# ---------------------------------------------------------------------------
# Feature Presets
# ---------------------------------------------------------------------------


class TestFeaturePresetsRoutes:
    def _create_yaml_file(self, path):
        """Create a minimal YAML file for preset tests."""
        import yaml

        os.makedirs(os.path.dirname(path), exist_ok=True)
        with open(path, "w") as f:
            yaml.dump({"test": True}, f)

    def test_get_feature_presets_empty(self, client, app, db_session):
        """GET /feature-presets should return a list (possibly empty)."""
        with _patch_validate_decorate(), _patch_decode_token():
            response = client.get("/feature-presets")

        assert response.status_code == 200
        data = response.get_json()
        assert isinstance(data, list)

    def test_get_feature_presets_with_data(self, client, app, db_session):
        """GET /feature-presets should return existing presets."""
        from quantimage2_backend_common.models import FeaturePreset

        yaml_path = "/tmp/quantimage2-test/route.yaml"
        self._create_yaml_file(yaml_path)

        with app.app_context():
            FeaturePreset("RoutePreset", yaml_path).save_to_db()

        with _patch_validate_decorate(), _patch_decode_token():
            response = client.get("/feature-presets")

        assert response.status_code == 200
        data = response.get_json()
        assert any(p.get("name") == "RoutePreset" for p in data)

    def test_get_single_feature_preset(self, client, app, db_session):
        """GET /feature-presets/<id> should return the preset."""
        from quantimage2_backend_common.models import FeaturePreset

        yaml_path = "/tmp/quantimage2-test/single.yaml"
        self._create_yaml_file(yaml_path)

        with app.app_context():
            preset = FeaturePreset("SinglePreset", yaml_path)
            preset.save_to_db()
            preset_id = preset.id

        with _patch_validate_decorate(), _patch_decode_token():
            response = client.get(f"/feature-presets/{preset_id}")

        assert response.status_code == 200
        data = response.get_json()
        assert data["name"] == "SinglePreset"


# ---------------------------------------------------------------------------
# Labels (LabelCategory)
# ---------------------------------------------------------------------------


class TestLabelsRoutes:
    def test_get_label_categories_for_album(self, client, app, db_session):
        """GET /label-categories/<album_id> should return label categories."""
        from quantimage2_backend_common.models import LabelCategory

        with app.app_context():
            LabelCategory(
                album_id="test-album",
                label_type="Classification",
                name="TestOutcome",
                user_id="test-user-uuid-1234",
            ).save_to_db()

        with _patch_validate_decorate(), _patch_decode_token():
            response = client.get("/label-categories/test-album")

        assert response.status_code == 200
        data = response.get_json()
        assert isinstance(data, list)


# ---------------------------------------------------------------------------
# Albums
# ---------------------------------------------------------------------------


class TestAlbumsRoutes:
    @patch("routes.albums.get_studies_from_album", return_value=[])
    def test_get_albums(self, mock_get_studies, client, app, db_session):
        """GET /albums/<album_id> should return album data."""
        with _patch_validate_decorate(), _patch_decode_token():
            response = client.get("/albums/test-album-route")

        # get_rois returns album.rois which is initially None (jsonified)
        assert response.status_code == 200


# ---------------------------------------------------------------------------
# Feature Extractions
# ---------------------------------------------------------------------------


class TestExtractionRoutes:
    def test_get_extraction_for_album(self, client, app, db_session):
        """GET /extractions/album/<album_id> should return the latest extraction."""
        from quantimage2_backend_common.models import FeatureExtraction

        with app.app_context():
            extraction = FeatureExtraction("test-user-uuid-1234", "test-extr-album")
            extraction.save_to_db()

        with _patch_validate_decorate(), _patch_decode_token():
            response = client.get("/extractions/album/test-extr-album")

        # May require Celery result mock — accept 200 or 500
        assert response.status_code in (200, 500)


# ---------------------------------------------------------------------------
# Feature Collections
# ---------------------------------------------------------------------------


class TestFeatureCollectionsRoutes:
    def test_get_collections_for_extraction(self, client, app, db_session):
        """GET /feature-collections/extraction/<extraction_id> should return collections."""
        from quantimage2_backend_common.models import (
            FeatureExtraction,
            FeatureCollection,
        )

        with app.app_context():
            extraction = FeatureExtraction("test-user-uuid-1234", "fc-album")
            extraction.save_to_db()
            extraction_id = extraction.id  # Capture before leaving context
            FeatureCollection(
                "TestColl",
                extraction_id,
                ["f1"],
                "fulldataset",
                "automatic",
                ["p1"],
                None,
            ).save_to_db()

        with _patch_validate_decorate(), _patch_decode_token():
            response = client.get(f"/feature-collections/extraction/{extraction_id}")

        assert response.status_code == 200
        data = response.get_json()
        assert isinstance(data, list)


# ---------------------------------------------------------------------------
# Models
# ---------------------------------------------------------------------------


class TestModelsRoutes:
    def test_get_models_for_album(self, client, app, db_session):
        """GET /models/<album_id> should return models for the album."""
        with _patch_validate_decorate(), _patch_decode_token():
            response = client.get("/models/test-model-album")

        assert response.status_code == 200
        data = response.get_json()
        assert isinstance(data, list)


# ---------------------------------------------------------------------------
# Navigation History
# ---------------------------------------------------------------------------


class TestNavigationRoutes:
    def test_post_navigation_history(self, client, app, db_session):
        """POST /navigation should create a history entry."""
        with _patch_validate_decorate(), _patch_decode_token():
            response = client.post(
                "/navigation",
                data=json.dumps({"path": "/albums/123"}),
                content_type="application/json",
            )

        assert response.status_code == 200
        data = response.get_json()
        assert data["path"] == "/albums/123"


# ---------------------------------------------------------------------------
# Error handling
# ---------------------------------------------------------------------------


class TestErrorHandling:
    def test_invalid_usage_returns_json_error(self, client, app):
        """InvalidUsage exceptions should return a JSON error response."""
        # The /test-error route is registered during app setup in conftest.py
        with _patch_validate_decorate(), _patch_decode_token():
            response = client.get("/test-error")

        assert response.status_code == 400
        data = response.get_json()
        assert data["message"] == "Test error message"


# ---------------------------------------------------------------------------
# Auth validation
# ---------------------------------------------------------------------------


class TestAuthValidation:
    def test_missing_auth_header_returns_error(self, app):
        """Requests without Authorization header should fail."""
        with app.test_client() as unauthed_client:
            response = unauthed_client.get("/feature-presets")

        # validate_decorate raises InvalidUsage or returns an error
        # Accept 400, 401, or 500 — the important thing is it's not 200
        assert response.status_code != 200
